Update from Master

+++ database/migrations/2015_04_12_093447_add_sv_language.php
@@ -12,7 +12,10 @@ class AddSvLanguage extends Migration {
 	 */
 	public function up()
 	{
-		DB::table('languages')->insert(['name' => 'Swedish', 'locale' => 'sv']);
+        DB::table('languages')->insert(['name' => 'Swedish', 'locale' => 'sv']);
+        DB::table('languages')->insert(['name' => 'Spanish - Spain', 'locale' => 'es_ES']);
+        DB::table('languages')->insert(['name' => 'French - Canada', 'locale' => 'fr_CA']);
+        DB::table('languages')->insert(['name' => 'Lithuanian', 'locale' => 'lt']);
 	}
 
 	/**
@@ -22,8 +25,21 @@ class AddSvLanguage extends Migration {
 	 */
 	public function down()
 	{
-        $language = \App\Models\Language::whereLocale('sv')->first();
-        $language->delete();
+        if ($language = \App\Models\Language::whereLocale('sv')->first()) {
+            $language->delete();
+        }
+
+        if ($language = \App\Models\Language::whereLocale('es_ES')->first()) {
+            $language->delete();
+        }
+
+        if ($language = \App\Models\Language::whereLocale('fr_CA')->first()) {
+            $language->delete();
+        }
+
+        if ($language = \App\Models\Language::whereLocale('lt')->first()) {
+            $language->delete();
+        }
 	}
 
 }
